feature: Creación entidad Articulo, Distribuidor y Carrito_Usuario

+++ MediExpressBack/MediExpress/src/main/java/ucentral/edu/co/MediExpress/Repositorios/UsuarioRepositorio.java
@@ -2,5 +2,5 @@ package ucentral.edu.co.MediExpress.Repositorios;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ucentral.edu.co.MediExpress.Entidades.Usuario;
 
-public interface UsuarioRepositorio extends JpaRepository<Usuario,Integer> {
+public interface UsuarioRepositorio extends JpaRepository<Usuario,Long> {
 }
